+++ python/lsst/ts/ofc/state_estimator.py
@@ -56,9 +56,7 @@ class StateEstimator:
         matrix.
     """
 
-    def __init__(
-        self, ofc_data: OFCData, rcond: float = 1e-3, log: logging.Logger | None = None
-    ) -> None:
+    def __init__(self, ofc_data: OFCData, log: logging.Logger | None = None) -> None:
         if log is None:
             self.log = logging.getLogger(type(self).__name__)
         else:
@@ -69,8 +67,10 @@ class StateEstimator:
         # Constuct the double zernike sensitivity matrix
         self.dz_sensitivity_matrix = SensitivityMatrix(self.ofc_data)
 
-        # Set rcond for the pseudoinverse computation
-        self.rcond = rcond
+        self.normalization_weights = ofc_data.normalization_weights
+
+        self.rcond = ofc_data.controller.get("truncation_threshold", None)
+        self.truncate_index = ofc_data.controller.get("truncation_index", None)
 
     def dof_state(
         self,
@@ -102,14 +102,22 @@ class StateEstimator:
 
         # Get the field angles for the sensors
         field_angles = [self.ofc_data.sample_points[sensor] for sensor in sensor_names]
+        field_x, field_y = zip(*field_angles)
+
+        rotation_angle_rad = np.deg2rad(-rotation_angle)
+        rot_mat = np.array(
+            [
+                [np.cos(rotation_angle_rad), -np.sin(rotation_angle_rad)],
+                [np.sin(rotation_angle_rad), np.cos(rotation_angle_rad)],
+            ]
+        )
+        field_angles = np.array([field_x, field_y]).T @ rot_mat
 
         # Evaluate sensitivity matrix at sensor positions
         # The rotation angle is negative because the rotation
         # in the uv-plane for the Double Zernike object
         # needs to be in the opposite direction.
-        sensitivity_matrix = self.dz_sensitivity_matrix.evaluate(
-            field_angles, -rotation_angle
-        )
+        sensitivity_matrix = self.dz_sensitivity_matrix.evaluate(field_angles, 0.0)
 
         # Select sensitivity matrix only at used degrees of freedom
         sensitivity_matrix = sensitivity_matrix[:, self.ofc_data.zn_idx, :]
@@ -122,6 +130,11 @@ class StateEstimator:
         # Select sensitivity matrix only at used degrees of freedom
         sensitivity_matrix = sensitivity_matrix[..., self.ofc_data.dof_idx]
 
+        normalization_matrix = np.diag(
+            self.normalization_weights[self.ofc_data.dof_idx]
+        )
+        sensitivity_matrix = sensitivity_matrix @ normalization_matrix
+
         # Check the dimension of sensitivity matrix to see if we can invert it
         num_zk, num_dof = sensitivity_matrix.shape
         if num_zk < num_dof:
@@ -131,6 +144,21 @@ class StateEstimator:
 
         # Compute the pseudo-inverse of the sensitivity matrix
         # rcond sets the truncation of different modes.
+        # If rcond is None, it is computed from the singular values
+        # of the sensitivity matrix, using the truncation index as reference.
+        if self.rcond is None and self.truncate_index is None:
+            raise ValueError(
+                "Neither truncation index or threshold are set in the controller."
+            )
+
+        if self.truncate_index is not None:
+            self.log.info("Setting rcond value from truncation index.")
+            _, s, _ = np.linalg.svd(sensitivity_matrix, full_matrices=False)
+            if self.truncate_index >= len(s):
+                self.rcond = 0.99 * s[-1] / np.max(s)
+            else:
+                self.rcond = 0.99 * s[self.truncate_index - 1] / np.max(s)
+
         pinv_sensitivity_matrix = np.linalg.pinv(sensitivity_matrix, rcond=self.rcond)
 
         # Rotate the wavefront error to the same orientation as the
@@ -145,8 +173,8 @@ class StateEstimator:
 
             zk_galsim = galsim.zernike.Zernike(
                 wfe_sensor,
-                R_outer=self.ofc_data.config["obscuration"]["radius_outer"],
-                R_inner=self.ofc_data.config["obscuration"]["radius_inner"],
+                R_outer=self.ofc_data.config["pupil"]["radius_outer"],
+                R_inner=self.ofc_data.config["pupil"]["radius_inner"],
             )
             # Note that we need to remove the first 4 Zernike coefficients
             # since we padded the wfe with zeros for the 4 first Zernike
@@ -154,6 +182,7 @@ class StateEstimator:
                 self.ofc_data.znmin :
             ]
 
+        self.log.debug(f"Derotated wavefront error: {wfe}")
         # Compute wavefront error deviation from the intrinsic wavefront error
         # y = wfe - intrinsic_zk - y2_correction
         # y2_correction is a static correction for the
@@ -174,6 +203,9 @@ class StateEstimator:
         y = y.reshape(-1, 1)
 
         # Compute optical state estimate in the basis of DOF
-        x = pinv_sensitivity_matrix.dot(y)
+        # Because of normalization, we need to de-normalize the result
+        # to retrieve the actual DOF values in the original 50 dimensional
+        # basis. For more details, see equation (10) in arXiv:2406.04656.
+        x = normalization_matrix @ pinv_sensitivity_matrix.dot(y)
 
         return x.ravel()