+++ tests/test_state_estimator.py
@@ -23,6 +23,7 @@ import pathlib
 import unittest
 
 import numpy as np
+import yaml
 from lsst.ts.ofc import OFCData, SensitivityMatrix, StateEstimator
 
 
@@ -33,13 +34,20 @@ class TestStateEstimator(unittest.TestCase):
         """Set up the test case."""
         self.ofc_data = OFCData("lsst")
 
-        self.estimator = StateEstimator(self.ofc_data, rcond=1e-5)
+        self.estimator = StateEstimator(self.ofc_data)
+        self.estimator.rcond = 1e-7
 
         self.wfe = np.loadtxt(
             pathlib.Path(__file__).parent.absolute()
             / "testData"
             / "lsst_wfs_error_iter0.z4c"
         )
+
+        file_path = (
+            pathlib.Path(__file__).parent.absolute() / "testData" / "test_weights.yaml"
+        )
+        with open(file_path, "r") as fp:
+            self.normalization_weights = np.array(yaml.safe_load(fp))
 
         dofs = np.loadtxt(
             pathlib.Path(__file__).parent.absolute() / "testData" / "lsst_dof_iter0.txt"
@@ -135,7 +143,67 @@ class TestStateEstimator(unittest.TestCase):
         residual = self.mean_squared_residual(
             sensitivity_matrix @ self.dofs, sensitivity_matrix @ state
         )
-        assert residual < 1e-3
+        assert residual < 2e-3
+
+    def test_dof_state_raises_if_no_truncation_method(self) -> None:
+        """Test dof_state raises an error if both rcond
+        and truncate_index are None.
+        """
+        self.estimator.rcond = None
+        self.estimator.truncate_index = None
+
+        with self.assertRaises(ValueError):
+            self.estimator.dof_state(
+                "R", self.wfe, self.sensor_name_list, rotation_angle=0.0
+            )
+
+    def test_dof_state_with_truncation_index(self) -> None:
+        """Test the dof_state method."""
+        # Set truncation index to be high
+        # we expect this to correspond to a threshold close to 1e-7
+        self.estimator.rcond = None
+        self.estimator.truncate_index = 46
+
+        # Compute sensitivity matrix
+        sensitivity_matrix = self.compute_sensitivity_matrix(
+            self.field_angles, rotation_angle=0.0
+        )
+
+        # Compute optical state estimate
+        state = self.estimator.dof_state(
+            "R", self.wfe, self.sensor_name_list, rotation_angle=0.0
+        )
+
+        # Check number of degrees of freedom matches the specified
+        n_values = len(self.estimator.ofc_data.dof_idx)
+        self.assertEqual(len(state), n_values)
+
+        # Check derived wavefront from state estimate
+        # matches the one from original dofs
+        residual = self.mean_squared_residual(
+            sensitivity_matrix @ self.dofs, sensitivity_matrix @ state
+        )
+        assert residual < 2e-3
+
+    def test_dof_state_with_normalization_weights(self) -> None:
+        """Test the dof_state method when using normalization weights."""
+        self.estimator.normalization_weights = self.normalization_weights
+
+        sensitivity_matrix = self.compute_sensitivity_matrix(
+            self.field_angles, rotation_angle=0.0
+        )
+
+        state = self.estimator.dof_state(
+            "R", self.wfe, self.sensor_name_list, rotation_angle=0.0
+        )
+
+        n_values = len(self.estimator.ofc_data.dof_idx)
+        self.assertEqual(len(state), n_values)
+
+        residual = self.mean_squared_residual(
+            sensitivity_matrix @ self.dofs, sensitivity_matrix @ state
+        )
+        assert residual < 2e-3
 
     def test_dof_state_trim_zn_dof(self) -> None:
         """Test the dof_state method with trimmed
@@ -173,7 +241,7 @@ class TestStateEstimator(unittest.TestCase):
         residual = self.mean_squared_residual(
             sensitivity_matrix @ self.dofs, sensitivity_matrix[..., :n_values] @ state
         )
-        assert residual < 1e-2
+        assert residual < 2e-2
 
     def test_dof_state_not_enough_zk(self) -> None:
         """Test the dof_state method with not enough zernike indices."""